Work on fixing gunw tests
