+++ tests/conftest.py
@@ -1,4 +1,11 @@
 import os
+import pathlib
+
+import pytest
+from moto import mock_aws
+from moto.core import patch_client
+
+import aws
 
 
 def pytest_configure(config):
@@ -9,3 +16,21 @@ def pytest_configure(config):
     os.environ['AWS_SESSION_TOKEN'] = 'testing'
     os.environ['AWS_DEFAULT_REGION'] = 'us-west-2'
     os.environ['AWS_REGION'] = 'us-west-2'
+
+
+@pytest.fixture
+def gunw_data_path():
+    return pathlib.Path(__file__).parent / 'data' / 'output_files' / 'gunw'
+
+
+@pytest.fixture
+def s3_bucket():
+    with mock_aws():
+        patch_client(aws.S3_CLIENT)
+
+        bucket_name = 'myBucket'
+        location = {'LocationConstraint': 'us-west-2'}
+
+        aws.S3_CLIENT.create_bucket(Bucket=bucket_name, CreateBucketConfiguration=location)
+
+        yield bucket_name
