+++ tests/test_aws.py
@@ -1,9 +1,3 @@
-import pathlib
-
-import pytest
-from moto import mock_aws
-from moto.core import patch_client
-
 import aws
 
 
@@ -25,21 +19,3 @@ def test_md5_for_s3_file_metadata(s3_bucket, gunw_data_path):
     metadata_md5 = '3b938e3797b8d5a90728ff64f7209752'
     assert aws.md5_for_s3_file(s3_bucket, metadata_key) == metadata_md5
     assert aws.md5_for_s3_file(s3_bucket, metadata_key, chunk_size=1024) == metadata_md5
-
-
-@pytest.fixture
-def gunw_data_path():
-    return pathlib.Path(__file__).parent / 'data' / 'output_files' / 'gunw'
-
-
-@pytest.fixture
-def s3_bucket():
-    with mock_aws():
-        patch_client(aws.S3_CLIENT)
-
-        bucket_name = 'myBucket'
-        location = {'LocationConstraint': 'us-west-2'}
-
-        aws.S3_CLIENT.create_bucket(Bucket=bucket_name, CreateBucketConfiguration=location)
-
-        yield bucket_name
