+++ tests/test_opera_rtc_s1_slc.py
@@ -157,15 +157,6 @@ def test_get_products(s3_stubber):
     ]
 
 
-def test_get_file_type():
-    assert opera_rtc_s1_slc._get_file_type('foo.tif') == 'data'
-    assert opera_rtc_s1_slc._get_file_type('bar.h5') == 'data'
-    assert opera_rtc_s1_slc._get_file_type('hello/world.iso.xml') == 'metadata'
-    assert opera_rtc_s1_slc._get_file_type('browse.png') == 'browse'
-    with pytest.raises(ValueError):
-        assert opera_rtc_s1_slc._get_file_type('bad_file.zip')
-
-
 def test_get_message(monkeypatch):
     now = datetime.datetime(2025, 2, 18, 1, 2, 3, 456, tzinfo=datetime.UTC)
     mock_datetime = MagicMock(wraps=datetime.datetime)
